Move business logic to `lib/`
As part of efforts to decouple the UI from the rest of the codebase,
move modules used by the pdf2md module, CLI and web UI into `lib/`.
Common practice dictates that the directory should be reserved for
code that can be used immediately, ie, not Bablified, but we will
ignore that convention for now so that we don't do too many things
in one go

[moves]
`src/javascript/models/AppState.jsx -> src/javascript/AppState.jsx`
`src/javascript/models -> lib/models`, tidying up subdir names
`src/javascript/lib -> lib/util`
`src/javascript/pdf2md*.js -> lib/pdf2md*.js`

[changes]
- point component modules and `npm run prepare` to `lib/`
- re-org `test/` to better align with `lib/`
