+++ lib/models/LineConverter.jsx
@@ -5,8 +5,8 @@ import WordFormat from './markdown/WordFormat.jsx';
 import LineItem from './LineItem.jsx';
 import StashingStream from './StashingStream.jsx';
 import { ParsedElements } from './PageItem.jsx';
-import { isNumber, isListItemCharacter } from '../lib/stringFunctions.jsx'
-import { sortByX } from '../lib/pageItemFunctions.jsx'
+import { isNumber, isListItemCharacter } from '../util/string-functions.jsx'
+import { sortByX } from '../util/page-item-functions.jsx'
 
 // Converts text items which have been grouped to a line (through TextItemLineGrouper) to a single LineItem doing inline transformations like
 //'whitespace removal', bold/emphasis annotation, link-detection, etc..
