+++ lib/models/LineItem.jsx
