+++ lib/models/LineItemBlock.jsx
