+++ lib/models/StashingStream.jsx
