+++ lib/models/TextItem.jsx
