+++ lib/models/markdown/WordType.jsx
