+++ lib/models/transformations/line-item-block/GatherBlocks.jsx
@@ -2,7 +2,7 @@ import ToLineItemBlockTransformation from '../ToLineItemBlockTransformation.jsx'
 import ParseResult from '../../ParseResult.jsx';
 import LineItemBlock from '../../LineItemBlock.jsx';
 import { DETECTED_ANNOTATION } from '../../Annotation.jsx';
-import { minXFromPageItems } from '../../../lib/pageItemFunctions.jsx';
+import { minXFromPageItems } from '../../../util/page-item-functions.jsx';
 
 // Gathers lines to blocks
 export default class GatherBlocks extends ToLineItemBlockTransformation {
