+++ lib/models/transformations/line-item/CompactLines.jsx
