+++ lib/models/transformations/line-item/RemoveRepetitiveElements.jsx
@@ -2,7 +2,7 @@ import ToLineItemTransformation from '../ToLineItemTransformation.jsx';
 import ParseResult from '../../ParseResult.jsx';
 import { REMOVED_ANNOTATION } from '../../Annotation.jsx';
 
-import { isDigit } from '../../../lib/stringFunctions.jsx'
+import { isDigit } from '../../../util/string-functions.jsx'
 
 
 function hashCodeIgnoringSpacesAndNumbers(string) {
