+++ lib/models/transformations/line-item/VerticalToHorizontal.jsx
