+++ lib/models/transformations/text-item/CalculateGlobalStats.jsx
