+++ lib/pdf2md.js
@@ -1,7 +1,7 @@
-const { parse } = require('./lib/pdf.jsx')
-const { makeTransformations, transform } = require('./lib/transformations.jsx')
+const { parse } = require('./util/pdf.jsx')
+const { makeTransformations, transform } = require('./util/transformations.jsx')
 if (typeof document === 'undefined') {
-  require('./lib/domstubs.js').setStubs(global)
+  require('./util/dom-stubs').setStubs(global)
 }
 /**
  * Reads a
