+++ lib/util/dom-stubs.js
