+++ lib/util/pdf.jsx
