+++ lib/util/string-functions.jsx
