+++ lib/util/transformations.jsx
@@ -1,15 +1,15 @@
-import CalculateGlobalStats from '../models/transformations/textitem/CalculateGlobalStats.jsx'
+import CalculateGlobalStats from '../models/transformations/text-item/CalculateGlobalStats.jsx'
 
-import CompactLines from '../models/transformations/lineitem/CompactLines.jsx'
-import RemoveRepetitiveElements from '../models/transformations/lineitem/RemoveRepetitiveElements.jsx'
-import VerticalToHorizontal from '../models/transformations/lineitem/VerticalToHorizontal.jsx'
-import DetectTOC from '../models/transformations/lineitem/DetectTOC.jsx'
-import DetectListItems from '../models/transformations/lineitem/DetectListItems.jsx'
-import DetectHeaders from '../models/transformations/lineitem/DetectHeaders.jsx'
+import CompactLines from '../models/transformations/line-item/CompactLines.jsx'
+import RemoveRepetitiveElements from '../models/transformations/line-item/RemoveRepetitiveElements.jsx'
+import VerticalToHorizontal from '../models/transformations/line-item/VerticalToHorizontal.jsx'
+import DetectTOC from '../models/transformations/line-item/DetectTOC.jsx'
+import DetectListItems from '../models/transformations/line-item/DetectListItems.jsx'
+import DetectHeaders from '../models/transformations/line-item/DetectHeaders.jsx'
 
-import GatherBlocks from '../models/transformations/textitemblock/GatherBlocks.jsx'
-import DetectCodeQuoteBlocks from '../models/transformations/textitemblock/DetectCodeQuoteBlocks.jsx'
-import DetectListLevels from '../models/transformations/textitemblock/DetectListLevels.jsx'
+import GatherBlocks from '../models/transformations/line-item-block/GatherBlocks.jsx'
+import DetectCodeQuoteBlocks from '../models/transformations/line-item-block/DetectCodeQuoteBlocks.jsx'
+import DetectListLevels from '../models/transformations/line-item-block/DetectListLevels.jsx'
 import ToTextBlocks from '../models/transformations/ToTextBlocks.jsx'
 import ToMarkdown from '../models/transformations/ToMarkdown.jsx'
