+++ package.json
@@ -15,7 +15,7 @@
     "check": "npm run lint && npm run test",
     "release": "npm run lint && rm -rf build/* && NODE_ENV=production webpack -p",
     "deploy": "npm run release && cp -r build/* docs/",
-    "prepare": "NODE_ENV=cli babel src/javascript --copy-files --out-dir dist/ --keep-file-extension"
+    "prepare": "NODE_ENV=cli babel lib --copy-files --out-dir dist/ --keep-file-extension"
   },
   "keywords": [
     "PDF",
