+++ src/javascript/index.jsx
@@ -5,7 +5,7 @@ import 'bootstrap/dist/css/bootstrap.css';
 import '../css/styles.css';
 
 import App from './components/App.jsx';
-import AppState from './models/AppState.jsx';
+import AppState from './AppState.jsx';
 
 function render(appState) {
     ReactDOM.render(<App appState={ appState } />, document.getElementById('main'));
